work/requirements.txt
torch
torchvision
pillow
numpy
flask